sys-libs/glibc: only copy crypt.h if it exists
Some folks may have rather strict INSTALL_MASKs and such
and don't need headers, so if they don't have it already,
go with it.

Bug: https://bugs.gentoo.org/802207

+++ sys-libs/glibc/glibc-9999.ebuild
@@ -1519,7 +1519,12 @@ pkg_preinst() {
 	if ! use crypt && has_version "${CATEGORY}/${PN}[crypt]"; then
 		PRESERVED_OLD_LIBCRYPT=1
 		preserve_old_lib /$(get_libdir)/libcrypt$(get_libname 1)
-		cp "${EROOT}"/usr/include/crypt.h "${T}"/crypt.h || die
+
+		# Only copy if it exists; some people may have tiny embedded
+		# systems without headers: https://bugs.gentoo.org/802207#c16
+		if [[ -f "${EROOT}"/usr/include/crypt.h ]] ; then
+			cp "${EROOT}"/usr/include/crypt.h "${T}"/crypt.h || die
+		fi
 	else
 		PRESERVED_OLD_LIBCRYPT=0
 	fi
@@ -1555,7 +1560,13 @@ pkg_postinst() {
 
 	if [[ ${PRESERVED_OLD_LIBCRYPT} -eq 1 ]] ; then
 		preserve_old_lib_notify /$(get_libdir)/libcrypt$(get_libname 1)
-		cp "${T}"/crypt.h "${EROOT}"/usr/include/crypt.h || eerror "Error restoring crypt.h, please file a bug"
+
+		# Only copy if it exists; some people may have tiny embedded
+		# systems without headers: https://bugs.gentoo.org/802207#c16
+		if [[ -f "${T}"/crypt.h ]] ; then
+			cp "${T}"/crypt.h "${EROOT}"/usr/include/crypt.h || eerror "Error restoring crypt.h, please file a bug"
+		fi
+
 		elog "Please ignore a possible later error message about a file collision involving"
 		elog "/usr/include/crypt.h. We need to preserve this file for the moment to keep"
 		elog "the upgrade working, but it also needs to be overwritten when"
